+++ lib/open_feature.rb
@@ -19,17 +19,33 @@ module OpenFeature
 
     sig { params(provider: Provider).void }
     def set_provider(provider) # rubocop:disable Naming/AccessorMethodName
-      configuration.set_provider(provider)
+      configuration.provider = provider
+    end
+
+    sig { params(context: EvaluationContext).void }
+    def set_evaluation_context(context) # rubocop:disable Naming/AccessorMethodName
+      configuration.evaluation_context = context
     end
 
     sig { params(hooks: T.any(Hook, T::Array[Hook])).void }
     def add_hooks(hooks)
-      configuration.add_hooks(hooks)
+      configuration.hooks.concat(Array(hooks))
     end
 
-    sig { params(name: T.nilable(String)).returns(Client) }
-    def create_client(name: nil)
-      Client.new(provider: configuration.provider, name: name)
+    sig do
+      params(
+        name: T.nilable(String),
+        evaluation_context: T.nilable(EvaluationContext),
+        hooks: T.nilable(T.any(Hook, T::Array[Hook]))
+      ).returns(Client)
+    end
+    def create_client(name: nil, evaluation_context: nil, hooks: nil)
+      Client.new(
+        provider: configuration.provider,
+        name: name,
+        evaluation_context: evaluation_context,
+        hooks: Array(hooks)
+      )
     end
 
     sig { returns(Configuration) }