+++ lib/open_feature/configuration.rb
@@ -13,14 +13,18 @@ module OpenFeature
     include Singleton
 
     sig { returns(Provider) }
-    attr_reader :provider
+    attr_accessor :provider
+
+    sig { returns(T.nilable(EvaluationContext)) }
+    attr_accessor :evaluation_context
 
     sig { returns(T::Array[Hook]) }
-    attr_reader :hooks
+    attr_accessor :hooks
 
     sig { void }
     def initialize
       @provider = T.let(NoOpProvider.new, Provider)
+      @evaluation_context = T.let(nil, T.nilable(EvaluationContext))
       @hooks = T.let([], T::Array[Hook])
     end
 
@@ -29,19 +33,11 @@ module OpenFeature
       provider.metadata
     end
 
-    sig { params(provider: Provider).void }
-    def set_provider(provider) # rubocop:disable Naming/AccessorMethodName
-      @provider = provider
-    end
-
-    sig { params(hooks: T.any(Hook, T::Array[Hook])).void }
-    def add_hooks(hooks)
-      @hooks.concat(Array(hooks))
-    end
-
     sig { void }
-    def clear_hooks!
+    def reset!
+      @provider = OpenFeature::NoOpProvider.new
       @hooks = []
+      @evaluation_context = nil
     end
   end
 end