+++ test/open_feature/configuration_test.rb
@@ -2,37 +2,34 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require_relative "../support/configuration_helper"
 require_relative "../support/test_provider"
 
 class ConfigurationTest < Minitest::Test
   def teardown
-    ConfigurationHelper.reset!
+    OpenFeature.configuration.reset!
   end
 
   def test_configuration_is_initialized_properly
     assert_equal("No Op Provider", OpenFeature::Configuration.instance.provider_metadata.name)
     assert_empty(OpenFeature::Configuration.instance.hooks)
+    assert_nil(OpenFeature::Configuration.instance.evaluation_context)
   end
 
-  def test_provider_can_be_set
-    OpenFeature.set_provider(TestProvider.new)
+  # rubocop:disable Metrics/AbcSize, Minitest/MultipleAssertions
+  def test_configuration_can_be_reset
+    OpenFeature::Configuration.instance.provider = TestProvider.new
+    OpenFeature::Configuration.instance.hooks = [OpenFeature::Hook.new, OpenFeature::Hook.new]
+    OpenFeature::Configuration.instance.evaluation_context = OpenFeature::EvaluationContext.new
 
     assert_equal("Test Provider", OpenFeature::Configuration.instance.provider_metadata.name)
-  end
+    assert_equal(2, OpenFeature::Configuration.instance.hooks.size)
+    refute_nil(OpenFeature::Configuration.instance.evaluation_context)
 
-  def test_hooks_can_be_added
-    OpenFeature::Configuration.instance.add_hooks(OpenFeature::Hook.new)
-    OpenFeature::Configuration.instance.add_hooks([OpenFeature::Hook.new, OpenFeature::Hook.new])
+    OpenFeature::Configuration.instance.reset!
 
-    assert_equal(3, OpenFeature::Configuration.instance.hooks.size)
-  end
-
-  def test_hooks_can_be_cleared
-    OpenFeature::Configuration.instance.add_hooks([OpenFeature::Hook.new, OpenFeature::Hook.new])
-
-    OpenFeature::Configuration.instance.clear_hooks!
-
+    assert_equal("No Op Provider", OpenFeature::Configuration.instance.provider_metadata.name)
     assert_empty(OpenFeature::Configuration.instance.hooks)
+    assert_nil(OpenFeature::Configuration.instance.evaluation_context)
   end
+  # rubocop:enable Metrics/AbcSize, Minitest/MultipleAssertions
 end