+++ test/support/configuration_helper.rb
@@ -1,9 +1,0 @@
-# typed: true
-# frozen_string_literal: true
-
-module ConfigurationHelper
-  def self.reset!
-    OpenFeature.set_provider(OpenFeature::NoOpProvider.new)
-    OpenFeature.configuration.clear_hooks!
-  end
-end