+++ CHANGELOG.md
@@ -6,6 +6,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-05-17
+
+### Added
+
+- Added ability to set evaluation context globally on the `Configuration` singleton, i.e. `OpenFeature::Configuration.instance.evaluation_context = OpenFeature::EvaluationContext.new(fields: { "globally" => "available" })`.
+- Added ability to set evaluation context globally on the `OpenFeature` module, i.e. `OpenFeature.set_evaluation_context(OpenFeature::EvaluationContext.new(fields: { "globally" => "available" }))`.
+- Added ability to set evaluation context on a `Client` instance, i.e. `client.evaluation_context = OpenFeature::EvaluationContext.new(fields: { "client" => "available" })`.
+- Added ability to set hooks and evaluation context on `Client` initialization, i.e. `OpenFeature.create_client(name: "my_client", evaluation_context: OpenFeature::EvaluationContext.new(fields: { "client" => "available" }), hooks: OpenFeature::Hook.new)`
+- Added `Configuration#reset!` to reset global configuration to the default state.
+
+### Changed
+
+- Renamed `Configuration#set_provider` to `Configuration#provider=`.
+- Renamed `Configuration#add_provider` to `Configuration#hooks=`.
+- During flag initialization, contexts are now merged. The invocation context takes precedence over the client context which takes precedence over the global context.
+
+### Removed
+
+- Removed `Configuration#clear_hooks!` in favor of `Configuration#reset!`.
+
 ## [0.1.2] - 2023-05-16
 
 ### Added